+++ ui/theming/ui/src/commonMain/kotlin/app/campfire/ui/theming/ui/builder/ThemeBuilderUi.kt
@@ -52,7 +52,6 @@ import app.campfire.common.compose.CampfireWindowInsets
 import app.campfire.common.compose.LocalWindowSizeClass
 import app.campfire.common.compose.currentWindowSizeClass
 import app.campfire.common.compose.icons.CampfireIcons
-import app.campfire.common.compose.icons.rounded.ArrowBack
 import app.campfire.common.compose.icons.rounded.DarkMode
 import app.campfire.common.compose.icons.rounded.Delete
 import app.campfire.common.compose.icons.rounded.LightMode
@@ -63,6 +62,7 @@ import app.campfire.common.compose.theme.LocalUseDarkColors
 import app.campfire.common.compose.theme.alt.AltRedColorPalette
 import app.campfire.common.compose.widgets.CampfireTopAppBar
 import app.campfire.common.compose.widgets.IconButtonTooltip
+import app.campfire.common.compose.widgets.NavigationBackButton
 import app.campfire.core.di.UserScope
 import app.campfire.ui.theming.api.AppTheme
 import app.campfire.ui.theming.api.colorScheme
@@ -74,7 +74,6 @@ import app.campfire.ui.theming.ui.builder.composables.ContrastPicker
 import app.campfire.ui.theming.ui.builder.composables.Header
 import app.campfire.ui.theming.ui.builder.composables.IconPicker
 import campfire.ui.theming.ui.generated.resources.Res
-import campfire.ui.theming.ui.generated.resources.action_back
 import campfire.ui.theming.ui.generated.resources.action_delete_theme
 import com.r0adkll.kimchi.circuit.annotations.CircuitInject
 import com.r0adkll.swatchbuckler.color.dynamiccolor.ColorSpec.SpecVersion
@@ -101,14 +100,7 @@ fun ThemeBuilder(
             Text("Theme Builder")
           },
           navigationIcon = {
-            val backLabel = stringResource(Res.string.action_back)
-            IconButtonTooltip(text = backLabel) {
-              IconButton(
-                onClick = { state.eventSink(ThemeBuilderUiEvent.Back) },
-              ) {
-                Icon(CampfireIcons.Rounded.ArrowBack, contentDescription = backLabel)
-              }
-            }
+            NavigationBackButton(onClick = { state.eventSink(ThemeBuilderUiEvent.Back) })
           },
           actions = {
             var showDeleteConfirmation by remember { mutableStateOf(false) }